Ajuste en BD y estructura de proyecto
